# mimbackend/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /auth/facebook/callback:
    get:
      description: Handles OAuth callback from Facebook
      parameters:
      - description: State
        in: query
        name: state
        required: true
        type: string
      - description: Code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Facebook OAuth callback
      tags:
      - Auth
...
  /auth/github/callback:
    get:
      description: Handles OAuth callback from Github
      parameters:
      - description: State
        in: query
        name: state
        required: true
        type: string
      - description: Code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Github OAuth callback
      tags:
      - Auth
...
  /auth/google/callback:
    get:
      description: Handles OAuth callback from Google
      parameters:
      - description: State
        in: query
        name: state
        required: true
        type: string
      - description: Code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Google OAuth callback
      tags:
      - Auth
...
swagger: "2.0"